refactor: Fix typo in background image URL
This commit fixes a typo in the background image URL for the "backdrop-padrinho" section in the index.html file. The image file name was misspelled as "bg_padriho.png" instead of "bg_padrinho.png". This correction ensures that the correct image is displayed, improving the visual consistency of the section.